--- save-cloud-common/src/commonMain/kotlin/com/saveourtool/save/execution/ExecutionType.kt
@@ -1,25 +1,0 @@
-package com.saveourtool.save.execution
-
-import kotlinx.serialization.Serializable
-
-/**
- * Type oj execution
- */
-@Serializable
-enum class ExecutionType {
-    /**
-     * Project on contest
-     */
-    CONTEST,
-
-    /**
-     * Project from git
-     */
-    GIT,
-
-    /**
-     * Project by binary file
-     */
-    STANDARD,
-    ;
-}
