+++ save-cloud-common/src/commonMain/kotlin/com/saveourtool/save/execution/TestingType.kt
@@ -1,0 +1,14 @@
+package com.saveourtool.save.execution
+
+import kotlinx.serialization.Serializable
+
+/**
+ * Types of testing (that can be selected by user)
+ */
+@Serializable
+enum class TestingType {
+    CONTEST_MODE,
+    PRIVATE_TESTS,
+    PUBLIC_TESTS,
+    ;
+}
