+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/basic/TestResourcesSelection.kt
@@ -7,6 +7,7 @@
 package com.saveourtool.save.frontend.components.basic
 
 import com.saveourtool.save.entities.ContestDto
+import com.saveourtool.save.execution.TestingType
 import com.saveourtool.save.frontend.components.basic.testsuiteselector.showPrivateTestSuitesSelectorModal
 import com.saveourtool.save.frontend.components.basic.testsuiteselector.showPublicTestSuitesSelectorModal
 import com.saveourtool.save.frontend.components.inputform.InputTypes
@@ -26,15 +27,6 @@ import react.dom.html.ReactHTML.select
 
 val testResourcesSelection = prepareTestResourcesSelection()
 
-/**
- * Types of testing (that can be selected by user)
- */
-enum class TestingType {
-    CONTEST_MODE,
-    PRIVATE_TESTS,
-    PUBLIC_TESTS,
-    ;
-}
 
 /**
  * Properties for test resources
